requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
appscript==1.3.0
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
beautifulsoup4==4.13.3
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cssselect2==0.7.0
decorator==4.4.2
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.19.1
fonttools==4.55.3
fpdf==1.7.2
frozenlist==1.4.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gspread==5.6.2
httplib2==0.22.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.5
lxml==5.3.0
magic-filter==1.0.12
MarkupSafe==3.0.2
moviepy==1.0.3
multidict==6.1.0
numpy==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
pdf2image==1.16.3
pdfkit==1.0.0
pillow==10.4.0
proglog==0.1.10
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
pypandoc==1.14
pyparsing==3.1.4
pyphen==0.17.0
python-docx==1.1.2
python-dotenv==0.21.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.6
SQLAlchemy==1.4.46
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
weasyprint==63.1
webencodings==0.5.1
yarl==1.14.0
zopfli==0.2.3.post1
